# Conflicts:
#	samples/GpuInterop/VulkanDemo/VulkanImage.cs
#	src/Avalonia.X11/X11Platform.cs
#	src/Windows/Avalonia.Win32/Win32GlManager.cs

+++ src/Avalonia.X11/X11PlatformThreading.cs
@@ -235,7 +235,7 @@ namespace Avalonia.X11
         }
 
 
-        public long Now => (int)_clock.ElapsedMilliseconds;
+        public long Now => _clock.ElapsedMilliseconds;
         public bool CanQueryPendingInput => true;
 
         public bool HasPendingInput => _platform.EventGrouperDispatchQueue.HasJobs || XPending(_display) != 0;
